pyproject.toml
[project]
name = "peelie-avatar-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "pgvector (>=0.4.2,<0.5.0)",
    "greenlet (>=3.3.2,<4.0.0)",
    "sqlalchemy (>=2.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "redis (>=7.3.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
minversion = "8.3.4"
env = []
addopts = " --strict-markers --tb=short --asyncio-mode=auto -p no:warnings --doctest-modules"
python_files = ["tests.py", "test_*.py"]
rootdir = "./"
testpaths = ["./tests"]
pythonpath = ["./"]
filterwarnings = ["error", "ignore::DeprecationWarning:etcd3.*:"]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
